+++ src/auth/certificates/__tests/VerifiableCertificate.test.ts
@@ -3,8 +3,9 @@
  * @description Tests for the VerifiableCertificate class.
  */
 
-import { VerifiableCertificate } from '../VerifiableCertificate'
-import { ProtoWallet, PrivateKey, SymmetricKey, Utils, Wallet, AbortActionArgs, AbortActionResult, AcquireCertificateArgs, AcquireCertificateResult, CreateActionArgs, CreateActionResult, DiscoverByAttributesArgs, DiscoverByIdentityKeyArgs, DiscoverCertificatesResult, GetHeaderArgs, GetHeaderResult, GetHeightResult, InternalizeActionArgs, InternalizeActionResult, ListActionsArgs, ListActionsResult, ListCertificatesArgs, ListCertificatesResult, ListOutputsArgs, ListOutputsResult, OriginatorDomainNameStringUnder250Bytes, ProveCertificateArgs, ProveCertificateResult, RelinquishCertificateArgs, RelinquishCertificateResult, RelinquishOutputArgs, RelinquishOutputResult, SignActionArgs, SignActionResult } from '../../../../mod'
+import { VerifiableCertificate } from '../../../../dist/cjs/src/auth/certificates/VerifiableCertificate.js'
+import { PrivateKey, SymmetricKey, Utils } from '../../../../dist/cjs/src/primitives/index.js'
+import { CompletedProtoWallet } from '../../../../dist/cjs/src/auth/certificates/__tests/MockWallet.js'
 
 describe('VerifiableCertificate', () => {
   const subjectPrivateKey = PrivateKey.fromRandom()
@@ -119,66 +120,3 @@ describe('VerifiableCertificate', () => {
     })
   })
 })
-
-class CompletedProtoWallet extends ProtoWallet implements Wallet {
-  async createAction(args: CreateActionArgs, originator?: OriginatorDomainNameStringUnder250Bytes)
-  : Promise<CreateActionResult> {
-    throw new Error("not implemented")
-  }
-  async signAction(args: SignActionArgs, originator?: OriginatorDomainNameStringUnder250Bytes)
-  : Promise<SignActionResult> {
-    throw new Error("not implemented")
-  }
-  async abortAction(args: AbortActionArgs, originator?: OriginatorDomainNameStringUnder250Bytes)
-  : Promise<AbortActionResult> {
-    throw new Error("not implemented")
-  }
-  async listActions(args: ListActionsArgs, originator?: OriginatorDomainNameStringUnder250Bytes)
-  : Promise<ListActionsResult> {
-    throw new Error("not implemented")
-  }
-  async internalizeAction(args: InternalizeActionArgs, originator?: OriginatorDomainNameStringUnder250Bytes)
-  : Promise<InternalizeActionResult> {
-    throw new Error("not implemented")
-  }
-  async listOutputs(args: ListOutputsArgs, originator?: OriginatorDomainNameStringUnder250Bytes)
-  : Promise<ListOutputsResult> {
-    throw new Error("not implemented")
-  }
-  async relinquishOutput(args: RelinquishOutputArgs, originator?: OriginatorDomainNameStringUnder250Bytes)
-  : Promise<RelinquishOutputResult> {
-    throw new Error("not implemented")
-  }
-  async acquireCertificate(args: AcquireCertificateArgs, originator?: OriginatorDomainNameStringUnder250Bytes)
-  : Promise<AcquireCertificateResult> {
-    throw new Error("not implemented")
-  }
-  async listCertificates(args: ListCertificatesArgs, originator?: OriginatorDomainNameStringUnder250Bytes)
-  : Promise<ListCertificatesResult> {
-    throw new Error("not implemented")
-  }
-  async proveCertificate(args: ProveCertificateArgs, originator?: OriginatorDomainNameStringUnder250Bytes)
-  : Promise<ProveCertificateResult> {
-    throw new Error("not implemented")
-  }
-  async relinquishCertificate(args: RelinquishCertificateArgs, originator?: OriginatorDomainNameStringUnder250Bytes)
-  : Promise<RelinquishCertificateResult> {
-    throw new Error("not implemented")
-  }
-  async discoverByIdentityKey(args: DiscoverByIdentityKeyArgs, originator?: OriginatorDomainNameStringUnder250Bytes)
-  : Promise<DiscoverCertificatesResult> {
-    throw new Error("not implemented")
-  }
-  async discoverByAttributes(args: DiscoverByAttributesArgs, originator?: OriginatorDomainNameStringUnder250Bytes)
-  : Promise<DiscoverCertificatesResult> {
-    throw new Error("not implemented")
-  }
-  async getHeight(args: {}, originator?: OriginatorDomainNameStringUnder250Bytes)
-  : Promise<GetHeightResult> {
-    throw new Error("not implemented")
-  }
-  async getHeaderForHeight(args: GetHeaderArgs, originator?: OriginatorDomainNameStringUnder250Bytes)
-  : Promise<GetHeaderResult> {
-    throw new Error("not implemented")
-  }
-}
