+++ src/auth/certificates/index.ts
@@ -1,3 +1,4 @@
 export { default as Certificate } from './Certificate.js'
 export * from './MasterCertificate.js'
 export * from './VerifiableCertificate.js'
+export * from './__tests/MockWallet.js'
